baby-care-ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.2.16
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langchain-ollama==0.1.3
langchain-openai==0.1.25
chromadb==0.4.24
faiss-cpu==1.7.4
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
PyYAML==6.0.1
numpy==1.24.3
sentence-transformers==5.0.0
requests==2.31.0
streamlit==1.28.1
streamlit-option-menu==0.3.6
plotly==5.17.0